Current Assessment / stds
a.       A back end or app repo with a Codacy badge - Yes.
b.       A back end or app  repo with a Snyk badge - Yes.
c.       A back end or app repo with the About section linked to their hosted site - Yes.
d.       A successfully deployed back end (or app) that is launched from the About section of their back end or app repo. - Yes.
e.       A working GitHub action that at a minimum runs prettier. - Yes.
x means a new pull request is required